Refactor entanglement generation protocols to use constants for formalism registration
